+++ src/main/java/com/raishxn/gtna/common/machine/multiblock/energy/NexusMEHyperCoreMachine.java
@@ -21,6 +21,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.block.Block;
 
+import com.raishxn.gtna.client.renderer.GTNATextures;
 import com.raishxn.gtna.common.data.GTNABlocks;
 import com.raishxn.gtna.common.machine.multiblock.part.ae.GTNACraftingCPUInterfacePartMachine;
 import com.raishxn.gtna.utils.Registries;
@@ -289,6 +290,8 @@ public class NexusMEHyperCoreMachine extends WorkableMultiblockMachine implement
         return new ModularUI(310, 270, this, entityPlayer)
                 .background(GuiTextures.BACKGROUND)
                 .widget(screen)
+                // The addon logo in the bottom-right corner of the content area (GTNL convention).
+                .widget(GTNATextures.logo(281, 161))
                 .widget(UITemplate.bindPlayerInventory(entityPlayer.getInventory(),
                         GuiTextures.SLOT, 74, 188, true));
     }